removed actions factories and define defaultActions in component itself
