--- plugins/catalog/src/components/CatalogTable/actions.tsx
@@ -1,67 +1,0 @@
-/*
- * Copyright 2021 Spotify AB
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-import React from 'react';
-import { Entity } from '@backstage/catalog-model';
-import {
-  getEntityMetadataViewUrl,
-  getEntityMetadataEditUrl,
-} from '@backstage/plugin-catalog-react';
-import OpenInNew from '@material-ui/icons/OpenInNew';
-import Edit from '@material-ui/icons/Edit';
-import {
-  favouriteEntityIcon,
-  favouriteEntityTooltip,
-} from '../FavouriteEntity/FavouriteEntity';
-
-export function createViewUrlAction(entity: Entity) {
-  const url = getEntityMetadataViewUrl(entity);
-  return {
-    icon: () => <OpenInNew fontSize="small" />,
-    tooltip: 'View',
-    disabled: !url,
-    onClick: () => {
-      if (!url) return;
-      window.open(url, '_blank');
-    },
-  };
-}
-
-export function createEditUrlAction(entity: Entity) {
-  const url = getEntityMetadataEditUrl(entity);
-  return {
-    icon: () => <Edit fontSize="small" />,
-    tooltip: 'Edit',
-    disabled: !url,
-    onClick: () => {
-      if (!url) return;
-      window.open(url, '_blank');
-    },
-  };
-}
-
-export function createStarredAction(
-  entity: Entity,
-  isStarredEntity: (entity: Entity) => boolean,
-  toggleStarredEntity: (entity: Entity) => void,
-) {
-  const isStarred = isStarredEntity(entity);
-  return {
-    cellStyle: { paddingLeft: '1em' },
-    icon: () => favouriteEntityIcon(isStarred),
-    tooltip: favouriteEntityTooltip(isStarred),
-    onClick: () => toggleStarredEntity(entity),
-  };
-}
